+++ example/lib/pages/location_stream_widget.dart
@@ -47,7 +47,7 @@ class LocationStreamState extends State<LocationStreamWidget> {
   @override
   Widget build(BuildContext context) {
     return FutureBuilder(
-        future: Geolocator.checkGeolocationStatus(),
+        future: Geolocator().checkGeolocationPermissionStatus(),
         builder:
             (BuildContext context, AsyncSnapshot<GeolocationStatus> snapshot) {
           if (!snapshot.hasData) {
@@ -121,8 +121,9 @@ class PositionListItemState extends State<PositionListItem> {
 
   void _onTap() async {
     String address = "unknown";
-    List<Placemark> placemarks = await Geolocator()
-        .placemarkFromCoordinates(_position.latitude, _position.longitude);
+    List<Placemark> placemarks = await Geolocator().placemarkFromCoordinates(
+        _position.latitude, _position.longitude,
+        localeIdentifier: "nl_NL");
 
     if (placemarks != null && placemarks.length > 0) {
       address = _buildAddressString(placemarks.first);